thêm chức năng, thêm sửa xoá truyện

+++ App/app/src/main/java/com/fpoly/assigment_mob403/DTO/Story.java
@@ -64,7 +64,7 @@ public class Story implements Serializable {
         return timeRelease;
     }
 
-    public void setTimeRelease(int timeRelease) {
+    public void setTimeRelease(long timeRelease) {
         this.timeRelease = timeRelease;
     }
